prevent uncustomizable components from being exported from index
