+++ src/components/AsideSection/AsideSection.tsx
@@ -1,33 +1,20 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 import '../../yamui';
 import * as React from 'react';
-import { NestableBaseComponentProps } from '../../util/BaseComponent/props';
-import { CustomizableComponentProps, nullTheme } from '../Customizer';
+import { CustomizableComponentProps, defaultTheme, customizable } from '../Customizer';
 import Block from '../Block';
 import { FixedGridColumn, FixedGridRow, GutterSize } from '../FixedGrid';
 import Heading from '../Heading';
+import { AsideSectionProps } from './AsideSection.types';
 import Text, { TextSize } from '../Text';
 import { getClassNames } from './AsideSection.styles';
-import { customizable } from '@uifabric/utilities';
-
-export interface AsideSectionProps extends NestableBaseComponentProps {
-  /**
-   * The visual title of the aside section
-   */
-  title: string;
-
-  /**
-   * An actionable react component to the right of the title
-   */
-  action?: React.ReactNode;
-}
 
 /**
  * A section component to be used primarily for sidebar modules
  */
 export class AsideSection extends React.Component<AsideSectionProps & CustomizableComponentProps> {
   public render() {
-    const { title, children, theme = nullTheme } = this.props;
+    const { title, children, theme = defaultTheme } = this.props;
     const classes = getClassNames({ theme });
 
     return (
