+++ src/components/AsideSection/index.ts
@@ -1,3 +1,3 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 export { default } from './AsideSection';
-export * from './AsideSection';
+export * from './AsideSection.types';
