+++ src/components/Avatar/Avatar.tsx
@@ -2,15 +2,11 @@
 import '../../yamui';
 import * as React from 'react';
 import { PersonaCoin, PersonaSize } from 'office-ui-fabric-react/lib/PersonaCoin';
-import { CustomizableComponentProps, nullTheme } from '../Customizer';
 import ScreenreaderText from '../ScreenreaderText';
-import { AvatarSize } from './types';
+import { AvatarSize, AvatarProps } from './Avatar.types';
 import { BorderType } from '../Image/types';
 import { getPersonaCoinStyles, getClassNames } from './Avatar.styles';
-import { customizable } from '@uifabric/utilities';
-import { ITheme } from '@uifabric/styling';
-
-export { BorderType, AvatarSize };
+import { defaultTheme, customizable } from '../Customizer';
 
 const SizeMap = {
   [AvatarSize.XLARGE]: PersonaSize.size72,
@@ -19,46 +15,6 @@ const SizeMap = {
   [AvatarSize.SMALL]: PersonaSize.size32,
   [AvatarSize.XSMALL]: PersonaSize.size24,
 };
-
-export interface AvatarProps extends CustomizableComponentProps {
-  /**
-   * Name of the person or object being represented. Will be used as accessible alt text.
-   */
-  name: string;
-
-  /**
-   * Element to be used as badge. You can set its height and width to fill the available area.
-   */
-  badgeContent?: React.ReactNode;
-
-  /**
-   * Short accessible description of the badge. Will be appended to name if provided.
-   */
-  badgeDescription?: string;
-
-  /**
-   * Type of border around the avatar.
-   * @default BorderType.ROUND
-   */
-  borderType?: BorderType;
-
-  /**
-   * Image source URL.
-   */
-  imageUrl?: string;
-
-  /**
-   * XLARGE: 72px, LARGE: 48px, MEDIUM: 40px, SMALL: 32px, XSMALL: 24px.
-   * @default AvatarSize.MEDIUM
-   */
-  size?: AvatarSize;
-
-  /**
-   * Will hide the image until it has loaded, then fade it in.
-   * @default false
-   */
-  imageShouldFadeIn?: boolean;
-}
 
 /**
  * An `Avatar` shows a thumbnail representation of both an individual or group.
@@ -70,12 +26,12 @@ export class Avatar extends React.Component<AvatarProps> {
       imageUrl,
       name,
       badgeDescription,
-      theme = nullTheme,
+      theme = defaultTheme,
       size = AvatarSize.MEDIUM,
       imageShouldFadeIn = false,
       borderType = BorderType.ROUND,
     } = this.props;
-    const classNames = getClassNames({ size, theme: theme as ITheme });
+    const classNames = getClassNames({ size, theme });
     const badge = badgeContent && <div className={classNames.badge}>{badgeContent}</div>;
 
     return (
