--- src/components/Avatar/types.ts
@@ -1,23 +1,0 @@
-/*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-export enum AvatarSize {
-  /**
-   * 72px.
-   */
-  XLARGE = 'xLarge',
-  /**
-   * 48px.
-   */
-  LARGE = 'large',
-  /**
-   * 40px;
-   */
-  MEDIUM = 'medium',
-  /**
-   * 32px.
-   */
-  SMALL = 'small',
-  /**
-   * 24px.
-   */
-  XSMALL = 'xSmall',
-}
