+++ src/components/Button/index.ts
@@ -1,6 +1,6 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
 export { default } from './Button';
-export * from './Button';
+export * from './Button.types';
 
 export { default as BistateButton } from './BistateButton';
 export * from './BistateButton';
